Switch back to non-custom image temporarily
Because of a bug with Docker Cloud automated builds it doesn't
seem possible to do the first `latest` build against a non-master
branch. So we need to get the CircleCI-based Dockerfile into
master in an earlier step than we start using the new image, or it
will be blocked from merging because the CircleCI build fails.