DON-746 – manually "cherry pick" eventual DON-204 solution
Meta Pixel seems to be functionally identical in all ways I can find to
2020's Facebook Pixel, including the URL to load it.

Apart from factoring in Meta's rebrand, the only change in this commit vs.
last time is to put the conversion tracking event at the success point in
`DonationStartComponent`, as we do with Analytics, since this has become
a safe (and more likely to load) option for us since our 2020 use of
FB Pixel, due to moving to only Stripe donations.