DON-715: Remove and .gitignore package-lock.json
angular/package-lock.json was outdated and caused a compile failure. It seems better
not to track it in git and make sure we update it as part of the build
process.

Having an outdated package-lock.json file has caused problems
a few times after a new component has been created - most recently
biggive-main-menu