DON-818: Copy popup modal modal from components library into this app, stop uising components version
Component code copied from here
https://github.com/thebiggive/components/tree/818f58fdf1bd3eec7d98f8d7d66a3313e4635fad/src/components/biggive-popup-standalone
where it can now be deleted, and just adapted slightly to work as an
Angular component

The reason is that there is an issue where text entered into the module
on Android phones (and perhaps other devices) appears backwards, because
somehow the curser keeps being pulled to the start of the input box as I
and others type.

I wasn't able to work out exactly what line of code in this component
causes the issue, but in my experiments using my phone yesterday I found
that moving the code into the login-modal.html file made the problem go
away, so I'm hoping this will also make it go away. Having the code here
will also make it much easier to experiemnt with changes in future.

I'm now in the office rather than at home, and with the different
network set up I'm not able to test my local envirionment with my phone,
so I think we should merge this to develop and test on staging.