Refactor: Required-input all the things
Taking advantage of the new feature in Angular 16 to express when an
input is always set. This gets checked at compile time in all templates
that use the component

See https://netbasal.com/from-good-to-great-required-inputs-in-angular-47374c8a43cf

I dif a fairly niave search for all the @Input decotators and added
required - then took it away anywhere it triggered a compile error.
Seems reasonable to assume that if we currently allways pass an input to
a componen then it's probably required.