DON-868: New stepper: fix keyboard navigation past gift aid radio controls
Previously I was able to tab into this radio button group, and tab out
of it backwards with shift-tab, but I wasn't able to tab out fowards.
Now I can properly use tab and shift-tab to move focus up and down the
page past the group, and the arrow keys to select an option.