DON-725: Remove min length validation from login modal
Not strictly related to the page, but I think this was confusing on the
modal. Better to let people try to log in with a short password if they
want to, and we can just tell them that their credentials are wrong.

Password min length only needs to be enforced at the time the password is
set.