DON-752: Adjust You Account page layout
Per comments on ticket.

I don't quite like the placement of "Log out", as it doesn't vertically
align with anything. I think eventually we might want to remove it from
here and put it in the main menu instead, but that should probably be
considered out of scope of the current ticket.