DON-950 - attempt to fix server side rendering
It seems like this makes the error I was seeing in server logs on my
local go away, but that may be a red herring as the homepage still
appears just as broken in the browser. Details below.

Error logged was "ERROR TypeError: Cannot read properties of undefined (reading 'apply')
    at /usr/src/app/dist/server/main.js:248821:65
    at _ZoneDelegate.invoke (/usr/src/app/dist/server/main.js:129405:160)
    at Zone.run (/usr/src/app/dist/server/main.js:129173:37)"

And from adding some debug code to main.js it seemed that the undefined
thing was `closeMobileMenuFromOutside`, hence adding the check for
isPlatformBrowser above that.