DON-965: Add cookie prefs links to footer & create dedicated page with cookie preference centre always open
On navigating to /cookie-preferences the preferences modal will be
automatically opened, and any time the user closes the model we navigate
away to /, so on that page the modal is always open.

The cookie preference modal has any preferences previously selected
pre-selected. So a new user will see all optional cookies turned off
by default, but a returning user who turned a cookie option on in the
past will see that option still turned on.