Fix ticker on fund pages
Throwing an error breaks the page, if there's no fund we don't want to
set the specific props yet, but we may need to set them slightly later.
Throwing an error broke that.

Changing type of setFundSpecificProps to
`(fund: Fund) => void` so we don't get errors
from calling it without a fund.