DON-713: Show highlight cards on explore page as well as homepage
But for now only showing if there are <= 6 cards and no search term
selected.

Not sure if we want to show this when there are search filters applied.

Also not sure we want to keep the existing "We can't find any campaigns
matching this search but there are lots more to choose from. View all
campaigns here."  text alongside this.