COM-99: Focus state on expansion panel & buttons on donate page
This is maybe expanding the scope of the ticket a bit, but we should
have consistent focus state indication across the page.

Adds a coral border around the expansion panel headed "How does Big Give
use tips?" and also the buttons on the page (mostly marked Continue)
when focused using the keyboard. Doesn't add if its focused using a
mouse or finger since that's generally unceassary.