BG2-2745: Allow registration via registration page without viewing donation funds form
By default after registering new donor account holders will now be
directed to the "my account" page.

I've kept the copy about transfering funds on the registration page but
only in the special case where a donor is on the way to view the
transfer funds form. It might be better to move this copy to the top of
that actual form - although people need to see it without logging in
so maybe even more ideally we'd allow at least part of the form to be
viewed without logging in. More than we want to change now.

I've also kept the "quick and easy" copy, although I'm not convinced its
helpful.