BG2-2772: Fix doulbe steper advance on payments click for logged in donor
As a logged in donor generally doesn't have to fill in anything in the
donation step (we pre-fill it with their saved info) its possible for
them to skip the step. The previous version accidentally made them skip
it by calling stepper.next() when they'd already told the stepper where
to go by clicking a header.