DON-1099: Update countries list used in Explore / Metacampaign to match SF
Our search API is in salesforce and works on matching exact names so we
don't find any results if these don't match.

The names in SF may not be all exactly how we would choose to spell
these countries but this is better than something that doesn't work.

SF also has Taiwan as "Chinese Taipei" but not changing that now since
our audience is less likely to know the latter name and we have no
campaigns operating there at present.