Merge meta-campaign component into explore
These two components look very similar and had lots of repeated code
between them. They both allowed exploring a set of campaigns, the
difference was that one was for exploring all BG campaigns in general
and the other was specific to campaigns within any given meta-campaign.

I went through and manually copied eveyry line of code from
meta-campaign to explore, adjust where necassary to deal with the fact
that having a meta-campaign defined is now optional