DON-1003: Display unmatched regular giving mandate
Changes to make the page at e.g.
/my-account/regular-giving/12ccf96e-b4d5-4c68-8963-b727f6abe212 work for
a mandate that does not use any match funds.

All copy adjusments are by me, need review.

The message
> Your donation of £12 is worth £12 to The Big Cat Sanctuary

feels a bit odd with the repeated same number, but it's the same as we
have on the ad-hoc donation thanks page, and it would be a different
number if gift aid was selected.

I don't know if we need to do anything more to make the difference
between matched and unmatched more obvious. Probably not on the full
page mandate view, but we might do on the list of mandates page