DON-1128: Start displaying mandate cancellation form while mandate loading
Previously none of the form displayed until the mandate was returned
from the server, meaning we only showed the header and footer of the
page in that time. Although I had anticipated that the component would
be delayed loading until the resolver resolves the mandate, that didn't
really happen as the resolver was only returning an observable of the
mandate, not the mandate itself.

There must be something built in to Angular that's implicitly unwrapping
the observable so we can use the mandate on this page.

I wasn't sure what the best way to fix this was and ended up being a bit
indecsive. The solution here is to display a lot more of the page before
the mandate is loaded, so that users get a better idea of what's
happening and can start to understand the new context while they wait
for loading to finish. If the mandate doesn't load within 1 second then
I also show a placeholder title and a spinner. I didn't want to go the
effort of building a full skelton UI.

Although this improves the Cancel Mandate page I think we still have the same
issue on the View Mandate page - so would be worth discussing how to
deal with this type of issue generally on the site.

Othere options I looked at were:

-  Making the observable return a promise of a mandate instead of an observable. This has the advantage of
fully stopping the loading of the component page until the mandate is
there, so we don't have to handle the case of it not being loaded within
the component. Would simplify things especially if we used that pattern
on several pages. But it has the problem that there is no visible
indication of loading - the URI doesn't update while waiting for the
resolve as we use the default 'deferred'  urlUpdateStrategy, not
'eager'. If we wanted to use this method I think we'd also need to
create a visible loading indication in the App component, maybe with
something like https://blog.angular-university.io/angular-loading-indicator/ . It might need to be built into
the main menu component.

- Caching the mandate so we don't have to wait for it to load. We might
  want to do something similar to how the CampaignResolver uses an
inject TransferState (although this wouldn't be transferred from the
server) to cache copies of Campaigns, or maybe something more like how
DonationService stores donation data in local storage (although I think
storage attached to the Window object is probably enough for mandates,
we don't to use Local Storage). This has a risk of showing outdated
mandate details if we don't get it quite right. I tried using a
record mapping mandate IDs observables with the shareReplay pipe in
the Mandate service to do something like this, but didn't quite get it
to work. But that is why links between two mandate pages now use
routerLink instead of href - href links destroy service state, which
routerLinks don't.