ID-47: Allow donor to view post-donation registration form
The plan is they will click a link in the donation thanks email (link
generation not yet implemented) to a URI that includes the six digit secret token
and their identity UUID, e.g.
/register?c=019816&u=1f0161fa-63ea-6fa8-888d-b1edd20e1c85

Donate then queries the identity backend for the details of the person,
and displays in the registration form as implemented in this commit.

The next step is to create a route in identity to allow registration to
be done using those details plus the newly chosen password, and
implement calling that route here.