Enable eslint with recomended Angular config
Linting was turned off in
https://github.com/thebiggive/donate-frontend/commit/ee72441b2368ae854c0b55ae15eda8637e9b70d2

Think we now have time to turn it back on.

For the moment I have switched off all rules that were generating errors
warnings. They can be switched back on and fixed one or a few at a time.

To install packages I just did `npm run lint` and followed the
suggestions from the CLI tool.