DON-521: Ensure every page has its own title (a11y improvement)
Previously a few pages (e.g. http://localhost:4200/transfer-funds and
http://localhost:4200/explore ) did not set specific titles so we just
showed 'Big Give' which is unhelpful in a list of tabs.

This commit makes sure every component has a title, and that the entry
in our routes array explicitly declares a title entry (although it may
be undefined) for every route.