DON-521: Add live announcment for screen readers when match funding expires
In testing whith the stock Ubuntu screen reader sometimes the appearence of the modal seemd to trigger an announcmenet without adding this,
and sometimes it didn't. Think it's worth having anyway.

There may be a wider redisgn of how we want to show match funding
expired anyway, which might mean moving away from a modal having an
aria-live region on the page instead into which we'd add the match
funding expired message when necassary - we could then block progress
through the form until the donor acknowledges the lack of match funding.

But I think this is good for now to make sure blind users know what the
modal dialog is about.

To test on my local I reduced the timer duration in my environment.ts to
0.3 minutes, and also commented out some code to make the timer run even
when there is zero match funds available.