DON-521: Fix links not keyboard focusable/clickable
I a couple of these and found it was a real issue that the link were not
focusable with the tab key and not clickable with enter key. Fixed here.

Not sure if there's any more direct way to tell angular to make an `a`
act like a link, I don't love the href attribute with no value but as we
have eslint to remind us to put it in where needed it's OK.

The original sin might be HTML using the same element for links and link
targets.