DON-521: Fix broken links introduced in recent accessability fix
Adding href to these links made them available to be used by keyboard,
but also broke them since the href didn't point to any sensible place.
Added a call to `Event.preventDefault()` for each one. Not sure if there
might be any more elegant way to do this but this does seem to work fine
when I test on local.

Also added a feature to dismiss the dubug info box on donation form,
e.g. in case it gets in the way when wanting to make a screenshot.
(Actually it was in adding that feature that I found all these links are
broken)