MAT-411: Display campaign errors on donation page
Should help us see what needs to be fixed to get matchbot campaign API
compatible as a replacement for salesforce API, including testing
against several different campaigns in staging.

Also moved the existing "debug info" developer panel on donate page to the sidebar
so it stops being in the way of the actual donation form. Maybe now
its dismiss function is unecassary but I've left it, I guess useful if
we want to take a screenshot without showing this.