DON-1147 – fix & use HMR, by narrowing browserslist for now
It seems like vite dev server falls over if we are asking for compatibility
with certain old browsers – and my first attempt to split development & production into
separate lists didn't work so I've done this for now.

Fortunately iOS and Safari ranges as we had before don't seem to cause the issue.