DON-1134: Improve UI for making unmatched regular giving donation
I think a checkbox inside a list doesn't really work well, so it makes
more sense to make the two checkboxes (aged over 18 and making unmatched
donation) sit together one after the other, and just use normal running
copy to describe the error.

Also making the error appear and disappear reactively when the donation
amount is reduced above or below the matchable maximum, and when
unmatched option is clicked.