DON-1152: Increase time to show errors for devs
Since (hopefully) every error will be unexepcted, I think it's worth
always displaying it for at least 5s, and up to 15 if its wordy.

We can still look in the console to see it again after.

Think we misinterpreted the code before and didn't realise it will by
default only show for 2s if the message is short.