DON-1142: Unseucssful attempt to remove optional chaining
Tried using Junie to remove optional chaining from all over the codebase
for compatibility with Chrome 79, but it makes it a lot more verbose and
also introduce lots of syntax errors.

Possibly for now we'll need to give up on Chrome 79 and show the
unsupported browser message for anything below Chrome 80.