DON-1142: Set up and eslint-plugin-compat to check we're not using new ES or web api features that our target browsers don't support.
It doesn't flag anything wrong in current code, but for instance if we
added a call to `new BroadcastChannel('x')` anywhere in our code this
would now be a build error because Safari 14 and 15 don't support it.

If we would need to use a BroadcasChannel (or anything else like that)
we could polyfill it and configure eslint-plugin-compat to know we have
a polyfill.