DON-1142 – commit possible index-varying ES5 fix
Largely the verbatim output of a couple of Copilot agents, following some back and forth for syntax corrections
etc. Probably a bit of a kitchen sink approach to varying what we serve currently – I feel like potentially we should
need only to vary the `<base/>` since individual injected scripts already do not separately include the /d/... path prefix.
But I think it probably makes sense to try this and see if it improves anything, then think about simplifications if it
is generally working.