DON-1156: Fix Swiper accessibility
See https://github.com/nolimits4web/swiper/issues/7528#issuecomment-3265739184

Worked around the swiper auto setting an ID that's inside the shadow
dom and putting that as the aria-controls attribute on the buttons by no
longer telling swiper about the buttons and instead telling the buttons
about the swiper, and manually setting the aria-controls attribute on
the buttons.

Also found that the only other thing stopping us getting a 100% score fo
accesibility with Chrome Lighthouse on the homepage was the swiper
pagination dots. These seemed to me to also be too small and low
contrast to be useful in any way for anyone, so I have just removed
them, and we now do get a 100% score.